Fix for compile with gcc 3.x, and compile on NetBSD and alpha

+++ src/io_helpers.cpp
@@ -299,7 +299,7 @@ size_t io::writeTrailingSpaces(ID3_Writer& writer, String buf, size_t len)
 {
   ID3_Writer::pos_type beg = writer.getCur();
   ID3_Writer::size_type strLen = buf.size();
-  ID3_Writer::size_type size = min(len,strLen);
+  ID3_Writer::size_type size = min((unsigned int)len, (unsigned int)strLen);
   writer.writeChars(buf.data(), size);
   for (; size < len; ++size)
   {
